YouTube Harvesting & Data Scraping
This project integrates with the YouTube Data API to fetch channel information, videos, playlists, and comments. It then stores this data into a PostgreSQL database for persistent storage and analysis. A Streamlit web application is provided to interact with the data, allowing users to execute predefined SQL queries and view results.